# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Antojados API
  version: 1.0.0
  description: API REST de AntojadosMX - Auth, Feed, Places, Posts, Social, Biz, Rewards, Media, Sync, Analytics, Equipo, GT Panel.
servers:
  - url: https://api.antojadosmx.mx
    description: Produccion (Cloudflare HTTPS)
  - url: http://185.187.235.253:8010
    description: Contabo (IP directa - fallback)
...
paths:
...
  /api/v1/antojados/auth/profile/{user_id}:
    get:
      tags: [Auth]
      summary: Obtener perfil
      parameters: [{in: path, name: user_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
    patch:
      tags: [Auth]
      summary: Actualizar perfil
      parameters: [{in: path, name: user_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/gt/moderation/{type}/{id}/approve:
    post:
      tags: [GT Moderation]
      summary: Aprobar item en moderacion
      parameters:
        - {in: path, name: type, required: true, schema: {type: string}}
        - {in: path, name: id, required: true, schema: {type: string}}
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/moderation/{type}/{id}/reject:
    post:
      tags: [GT Moderation]
      summary: Rechazar item en moderacion
      parameters:
        - {in: path, name: type, required: true, schema: {type: string}}
        - {in: path, name: id, required: true, schema: {type: string}}
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/gt/dimensions/{code}/status:
    patch:
      tags: [GT Dimensions]
      summary: Actualizar estado de dimension
      parameters: [{in: path, name: code, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/dimensions/{code}:
    delete:
      tags: [GT Dimensions]
      summary: Eliminar dimension
      parameters: [{in: path, name: code, required: true, schema: {type: string}}]
      responses:
        "204": { description: No Content }
...